backend/requirements.txt
Django==5.1.5
sqlparse==0.5.3
djangorestframework==3.15.2
markdown==3.7
django-filter==24.3
django-environ==0.12.0
openmeteo-requests==1.3.0
requests-cache==1.2.1
retry-requests==2.0.0
numpy==2.2.2
